refactor: changed casts to be more idiomatic

+++ src/ui/raw.rs
@@ -15,15 +15,15 @@ impl WidgetRef for RawMonoVisualizer {
 
         let canvas = Canvas::default()
             .x_bounds([0., (amps.len() - 1) as f64])
-            .y_bounds([area.top() as f64, area.bottom() as f64])
+            .y_bounds([f64::from(area.top()), f64::from(area.bottom())])
             .paint(|c| {
                 let mut j = 0.;
                 for points in amps.windows(2) {
                     c.draw(&Line {
                         x1: j,
-                        y1: mid + points[0] as f64,
+                        y1: mid + points[0],
                         x2: j + 1.,
-                        y2: mid + points[1] as f64,
+                        y2: mid + points[1],
                         color: Color::Reset,
                     });
                     j += 1.;
